refactor: accountReducer 함수 분리
