+++ src/Reducer/index.js
@@ -1,0 +1,7 @@
+export {accountReducer} from './AccountRedcuer';
+export {
+  insertMoney,
+  refundMoney,
+  buyProduct,
+  inputMoney,
+} from './AccountActions';
